Plumb through cluster spec metadata storage with etcd snapshots Refactor init node election/search Organize functions into designated files for easier readability/maintenance
