+++ pkg/provisioningv2/rke2/planner/defaultplan.go
@@ -6,15 +6,16 @@ import (
 	rkev1 "github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1"
 	"github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1/plan"
 	"github.com/rancher/rancher/pkg/provisioningv2/rke2/runtime"
-	corecontrollers "github.com/rancher/wrangler/pkg/generated/controllers/core/v1"
 )
 
-func commonNodePlan(secrets corecontrollers.SecretCache, controlPlane *rkev1.RKEControlPlane, options plan.NodePlan) (plan.NodePlan, error) {
+// commonNodePlan returns a "default" node plan with the corresponding registry configuration.
+// It will append to the node plan passed in through options.
+func (p *Planner) commonNodePlan(controlPlane *rkev1.RKEControlPlane, options plan.NodePlan) (plan.NodePlan, error) {
 	if controlPlane.Spec.Registries == nil {
 		return options, nil
 	}
 
-	registryConfig, files, err := toRegistryConfig(secrets, runtime.GetRuntime(controlPlane.Spec.KubernetesVersion),
+	registryConfig, files, err := p.toRegistryConfig(runtime.GetRuntime(controlPlane.Spec.KubernetesVersion),
 		controlPlane.Namespace, controlPlane.Spec.Registries)
 	if err != nil {
 		return plan.NodePlan{}, err
