+++ pkg/provisioningv2/rke2/planner/etcdcreate.go
@@ -6,59 +6,37 @@ import (
 
 	rkev1 "github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1"
 	"github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1/plan"
-	rkecontroller "github.com/rancher/rancher/pkg/generated/controllers/rke.cattle.io/v1"
 	"github.com/rancher/rancher/pkg/provisioningv2/rke2/runtime"
-	"github.com/rancher/rancher/pkg/wrangler"
-	corecontrollers "github.com/rancher/wrangler/pkg/generated/controllers/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 )
 
-type etcdCreate struct {
-	controlPlane rkecontroller.RKEControlPlaneClient
-	secrets      corecontrollers.SecretCache
-	store        *PlanStore
-	s3Args       *s3Args
-}
-
-func newETCDCreate(clients *wrangler.Context, store *PlanStore) *etcdCreate {
-	return &etcdCreate{
-		controlPlane: clients.RKE.RKEControlPlane(),
-		secrets:      clients.Core.Secret().Cache(),
-		store:        store,
-		s3Args: &s3Args{
-			secretCache: clients.Core.Secret().Cache(),
-			env:         true,
-		},
-	}
-}
-
-func (e *etcdCreate) setState(controlPlane *rkev1.RKEControlPlane, spec *rkev1.ETCDSnapshotCreate, phase rkev1.ETCDSnapshotPhase) error {
+func (p *Planner) setEtcdSnapshotCreateState(controlPlane *rkev1.RKEControlPlane, spec *rkev1.ETCDSnapshotCreate, phase rkev1.ETCDSnapshotPhase) error {
 	controlPlane = controlPlane.DeepCopy()
 	controlPlane.Status.ETCDSnapshotCreatePhase = phase
 	controlPlane.Status.ETCDSnapshotCreate = spec
-	_, err := e.controlPlane.UpdateStatus(controlPlane)
+	_, err := p.rkeControlPlanes.UpdateStatus(controlPlane)
 	if err != nil {
 		return err
 	}
 	return ErrWaiting("refreshing etcd create state")
 }
 
-func (e *etcdCreate) resetEtcdCreateState(controlPlane *rkev1.RKEControlPlane) error {
+func (p *Planner) resetEtcdSnapshotCreateState(controlPlane *rkev1.RKEControlPlane) error {
 	if controlPlane.Status.ETCDSnapshotCreate == nil && controlPlane.Status.ETCDSnapshotCreatePhase == "" {
 		return nil
 	}
-	return e.setState(controlPlane, nil, "")
+	return p.setEtcdSnapshotCreateState(controlPlane, nil, "")
 }
 
-func (e *etcdCreate) startOrRestartCreate(controlPlane *rkev1.RKEControlPlane, snapshot *rkev1.ETCDSnapshotCreate) error {
+func (p *Planner) startOrRestartEtcdSnapshotCreate(controlPlane *rkev1.RKEControlPlane, snapshot *rkev1.ETCDSnapshotCreate) error {
 	if controlPlane.Status.ETCDSnapshotCreate == nil || !equality.Semantic.DeepEqual(*snapshot, *controlPlane.Status.ETCDSnapshotCreate) {
-		return e.setState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseStarted)
+		return p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseStarted)
 	}
 	return nil
 }
 
-func (e *etcdCreate) etcdCreate(controlPlane *rkev1.RKEControlPlane, clusterPlan *plan.Plan, snapshot *rkev1.ETCDSnapshotCreate) []error {
+func (p *Planner) runEtcdSnapshotCreate(controlPlane *rkev1.RKEControlPlane, clusterPlan *plan.Plan, snapshot *rkev1.ETCDSnapshotCreate) []error {
 	servers := collect(clusterPlan, func(machine *capi.Machine) bool {
 		if !isEtcd(machine) || machine.Status.NodeRef == nil {
 			return false
@@ -74,18 +52,18 @@ func (e *etcdCreate) etcdCreate(controlPlane *rkev1.RKEControlPlane, clusterPlan
 	var errs []error
 
 	for _, server := range servers {
-		createPlan, err := e.createPlan(controlPlane, snapshot, server.Machine.Status.NodeRef.Name)
+		createPlan, err := p.generateEtcdSnapshotCreatePlan(controlPlane, snapshot, server.Machine.Status.NodeRef.Name)
 		if err != nil {
 			return []error{err}
 		}
-		if err := assignAndCheckPlan(e.store, "etcd snapshot", server, createPlan, 3); err != nil {
+		if err := assignAndCheckPlan(p.store, "etcd snapshot", server, createPlan, 3); err != nil {
 			errs = append(errs, err)
 		}
 	}
 	return errs
 }
 
-func (e *etcdCreate) createPlan(controlPlane *rkev1.RKEControlPlane, snapshot *rkev1.ETCDSnapshotCreate, nodeName string) (plan.NodePlan, error) {
+func (p *Planner) generateEtcdSnapshotCreatePlan(controlPlane *rkev1.RKEControlPlane, snapshot *rkev1.ETCDSnapshotCreate, nodeName string) (plan.NodePlan, error) {
 	args := []string{
 		"etcd-snapshot",
 	}
@@ -97,12 +75,12 @@ func (e *etcdCreate) createPlan(controlPlane *rkev1.RKEControlPlane, snapshot *r
 		args = append(args, fmt.Sprintf("--node-name=%s", nodeName))
 	}
 
-	s3Args, s3Env, s3Files, err := e.s3Args.ToArgs(snapshot.S3, controlPlane)
+	s3Args, s3Env, s3Files, err := p.etcdS3Args.ToArgs(snapshot.S3, controlPlane)
 	if err != nil {
 		return plan.NodePlan{}, err
 	}
 
-	return commonNodePlan(e.secrets, controlPlane, plan.NodePlan{
+	return p.commonNodePlan(controlPlane, plan.NodePlan{
 		Files: s3Files,
 		Instructions: []plan.Instruction{{
 			Name:    "create",
@@ -113,23 +91,21 @@ func (e *etcdCreate) createPlan(controlPlane *rkev1.RKEControlPlane, snapshot *r
 	})
 }
 
-func (e *etcdCreate) Create(controlPlane *rkev1.RKEControlPlane, clusterPlan *plan.Plan) []error {
+func (p *Planner) createEtcdSnapshot(controlPlane *rkev1.RKEControlPlane, clusterPlan *plan.Plan) []error {
 	if !Provisioned.IsTrue(controlPlane) && controlPlane.Status.ETCDSnapshotCreatePhase == "" {
 		return nil
 	}
 
 	if controlPlane.Spec.ETCDSnapshotCreate == nil {
-		if err := e.resetEtcdCreateState(controlPlane); err != nil {
+		if err := p.resetEtcdSnapshotCreateState(controlPlane); err != nil {
 			return []error{err}
 		}
 		return nil
 	}
 
-	return e.createSnapshot(controlPlane, clusterPlan, controlPlane.Spec.ETCDSnapshotCreate)
-}
+	snapshot := controlPlane.Spec.ETCDSnapshotCreate
 
-func (e *etcdCreate) createSnapshot(controlPlane *rkev1.RKEControlPlane, clusterPlan *plan.Plan, snapshot *rkev1.ETCDSnapshotCreate) []error {
-	if err := e.startOrRestartCreate(controlPlane, snapshot); err != nil {
+	if err := p.startOrRestartEtcdSnapshotCreate(controlPlane, snapshot); err != nil {
 		return []error{err}
 	}
 
@@ -137,7 +113,7 @@ func (e *etcdCreate) createSnapshot(controlPlane *rkev1.RKEControlPlane, cluster
 	case rkev1.ETCDSnapshotPhaseStarted:
 		var stateSet bool
 		var finErrs []error
-		if errs := e.etcdCreate(controlPlane, clusterPlan, snapshot); len(errs) > 0 {
+		if errs := p.runEtcdSnapshotCreate(controlPlane, clusterPlan, snapshot); len(errs) > 0 {
 			for _, err := range errs {
 				if err == nil {
 					continue
@@ -147,7 +123,7 @@ func (e *etcdCreate) createSnapshot(controlPlane *rkev1.RKEControlPlane, cluster
 				if !errors.As(err, &errWaiting) {
 					// we have a failed snapshot from a node.
 					if !stateSet {
-						if err := e.setState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseFailed); err != nil {
+						if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseFailed); err != nil {
 							finErrs = append(finErrs, err)
 						} else {
 							stateSet = true
@@ -157,7 +133,7 @@ func (e *etcdCreate) createSnapshot(controlPlane *rkev1.RKEControlPlane, cluster
 			}
 			return finErrs
 		}
-		if err := e.setState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseFinished); err != nil {
+		if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseFinished); err != nil {
 			return []error{err}
 		}
 		return nil
@@ -166,7 +142,7 @@ func (e *etcdCreate) createSnapshot(controlPlane *rkev1.RKEControlPlane, cluster
 	case rkev1.ETCDSnapshotPhaseFinished:
 		return nil
 	default:
-		if err := e.setState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseStarted); err != nil {
+		if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseStarted); err != nil {
 			return []error{err}
 		}
 		return nil
